refactor : Section 삭제 시, user 체크
- 임시 user_id를 받아와서 Service layer에서 course를 만든 사람이 맞는지 확인한다.

+++ src/main/java/com/example/projectlxp/section/service/impl/SectionServiceImpl.java
@@ -97,12 +97,17 @@ public class SectionServiceImpl implements SectionService {
 
     @Override
     @Transactional
-    public void removeSection(Long sectionId) {
+    public void removeSection(Long userId, Long sectionId) {
         // find section
         Section findSection =
                 sectionRepository
                         .findById(sectionId)
                         .orElseThrow(() -> new IllegalArgumentException("섹션이 존재하지 않습니다."));
+
+        // check who created this section
+        if (findSection.getCourse().getInstructor().getId() != userId) {
+            throw new IllegalArgumentException("섹션을 삭제할 권한이 없습니다.");
+        }
 
         // delete section
         sectionRepository.deleteById(sectionId);
